Abstracted cross product datatypes into new cross.py
